abi: share the typed syscall ABI (structs + flags), finishing the dedup
struct timeval/timespec/stat/dirent + tty_cell_t and the CLOCK_*, open/
fcntl flags and S_IF/DT_ bits were defined twice and 'must match by hand'
in <kernel/syscall.h> and src/userspace/syscall.h.  Hoist them into a
shared makar_abi.h (next to makar_syscalls/signals/keys), using plain
unsigned int/short/int -- layout-identical to the kernel's old uint32_t/
uint16_t on i386, and usable under the userspace -nostdinc build.

Kernel keeps FD_*/SYSCALL_FILE_* (internal); userspace keeps O_NONBLOCK,
access modes, S_ISxxx, sig_handler_t, SEEK/VGA/mmap (userspace-only).
Shipped to /usr/include for in-OS tcc.  Gates green: build, ktest 874,
iso test (incl in-OS libc-tcc), guitest.
